\name{NormPsy-package}
\alias{NormPsy-package}
\alias{NormPsy}
\docType{package}
\title{
Normalization of psychometric test scores
}
\description{
This package provides functions for normalizing psychometric test scores. The normalization aims at correcting the metrological properties of the psychometric tests such as the ceiling and floor effects and the curvilinearity (unequal interval scaling). The package also provides functions to compute and plot predictions in the natural scale of the psychometric test from the estimates of a linear mixed model estimated on the normalized scores.
}
\details{
\tabular{ll}{
Package: \tab NormPsy\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-05-07\cr
License: \tab GPL\cr
}
}
\author{
Cecile Proust-Lima, Viviane Philipps
}
%\references{
%}
